Lesson 6 fixed + a little fix Lesson 4
