+++ src/com/zadziarnouski/Lesson4/Task18.java
@@ -8,15 +8,13 @@ package com.zadziarnouski.Lesson4;
 
 public class Task18 {
     public static void main(String[] args) {
-        String str = "Стро;ковый литерал — это последова:тель-ность символов, закл+юченных в двойные кавычки.";
+        String str = "Стро;ковый литерал — это п!осл_едова:тель-ность символов, закл+юченных в двойные кавычки.";
         int counter = 0;
-        char symbol;
-        for (int i = 0; i < str.length(); i++) {
-            symbol = str.charAt(i);
-            if (symbol == '—' || symbol == ',' || symbol == '.' || symbol == '-' || symbol == ':' || symbol == ';' || symbol == '+') {
-                counter++;
-            }
+        for (String string : str.split("\\p{Punct}")) {
+            counter++;
         }
-        System.out.println(String.format("В этом предложении %d знаков припинания или др.символов", counter));
+        System.out.println(String.format("В данной строке %d знаков припинания.", counter));
+
+
     }
 }
