+++ src/com/zadziarnouski/Lesson6/Runner.java
@@ -4,8 +4,7 @@ public class Runner {
     public static void main(String[] args) {
         Laptop dell = new Laptop("i7", "16 GB", "1 TB");
         dell.turnOn(1);
-        dell.turnOff(0);
-        dell.turnOn(1);
         dell.turnOn(0);
+
     }
 }
